# laravel/openapi.yaml
openapi: 3.0.0
info:
  title: 'API Documentation'
  contact:
    email: info@mail.com
  version: '0.1'
servers:
  -
    url: 'https://localhost/api/'
    description: Server
paths:
...
  /auto_model:
    get:
      tags:
        - Auto
      operationId: 01fd57b871819e3a209e3c4e2646b43e
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AutoModelResource'
        '401':
          description: Unauthorized
        '404':
          description: 'Not Found'
    post:
      tags:
        - Auto
      operationId: 9be62fbbb2a9864655140374905b1281
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutoModelRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutoModelResource'
        '401':
          description: Unauthorized
        '404':
          description: 'Not Found'
  '/auto_model/{id}':
    delete:
      tags:
        - Auto
      operationId: 9bc0f999e51d1b11aab75135697cb1f3
      parameters:
        -
          name: id
          in: path
          description: 'ID of auto model'
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Ok
        '401':
          description: Unauthorized
        '404':
          description: 'Not Found'
    patch:
      tags:
        - Auto
      operationId: c6be81a9458127495efe0c6db8b3c7e3
      parameters:
        -
          name: id
          in: path
          description: 'ID of auto model'
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutoModelRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutoModelResource'
        '401':
          description: Unauthorized
        '404':
          description: 'Not Found'
  /auth/login:
    post:
      tags:
        - Authorization
      operationId: 6ac96e5dfa799cab4041324fc232ac42
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {  }
        '401':
          description: UNAUTHORIZED
components:
  schemas:
...
    AutoModelRequest:
      required:
        - name
      properties:
        name:
          $ref: '#/components/schemas/AutoModel/properties/name'
      type: object
...
    AutoModelResource:
      properties:
        id:
          $ref: '#/components/schemas/AutoModel/properties/id'
        name:
          $ref: '#/components/schemas/AutoModel/properties/name'
        created_at:
          $ref: '#/components/schemas/AutoModel/properties/created_at'
        updated_at:
          $ref: '#/components/schemas/AutoModel/properties/updated_at'
      type: object
...
    UserLoginRequest:
      required:
        - email
        - password
      properties:
        email:
          description: E-Mail
          type: string
          example: test@test.com
        password:
          description: 'User password'
          type: string
          example: '123456'
      type: object
...